% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstracts2text.R
\name{abstracts2text}
\alias{abstracts2text}
\title{Validate a data frame with abstracts to construct vectpr of strings}
\usage{
abstracts2text(abstracts = NULL)
}
\arguments{
\item{abstracts}{An object of the class abstracts (a data frame with abstracts) or
a string vector with abstracts}
}
\value{
A validated \code{character} vector. The function check and encode characters
in utf8 standard.
}
\description{
Validate a data frame with abstracts to construct vectpr of strings
}
\examples{
abstracts <- get_abstracts(lacsSample)
abstracts2text(abstracts)
}
