% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_abstract.R
\docType{data}
\name{example_abstract}
\alias{example_abstract}
\title{Vector with 100 examples of abstracts from papers. This abstracts are used
as an example to create a object of class abstracts.}
\format{
An object of class \code{"character"} with 100 abstracts.
}
\usage{
data(example_abstract)
}
\description{
Vector with 100 examples of abstracts from papers. This abstracts are used
as an example to create a object of class abstracts.
}
\examples{
data(example_abstract)
}
\keyword{datasets}
