% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_auc.R
\docType{data}
\name{example_auc}
\alias{example_auc}
\title{Example of value auc for test data using a plus model with a set of abstracts.}
\format{
An numeric value of auc;
}
\usage{
data(example_auc)
}
\description{
Example of value auc for test data using a plus model with a set of abstracts.
}
\examples{
data(example_auc)
}
\keyword{datasets}
