% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_class.R
\docType{data}
\name{example_class}
\alias{example_class}
\title{Vector with 100 examples of class of abstracts from papers. Is used as an example to build
an abstracts class. Could be positive (belonging to the corpus of mammal - parasite
papers) or unknown.}
\format{
An object of class \code{"character"} with 100 classes of abstracts.
}
\usage{
data(example_class)
}
\description{
Vector with 100 examples of class of abstracts from papers. Is used as an example to build
an abstracts class. Could be positive (belonging to the corpus of mammal - parasite
papers) or unknown.
}
\examples{
data(example_class)
}
\keyword{datasets}
