% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_doi.R
\docType{data}
\name{example_doi}
\alias{example_doi}
\title{Vector with 100 examples of DOI from papers. Is used as an example to build
an abstracts class.}
\format{
An object of class \code{"character"} with 100 DOI.
}
\usage{
data(example_doi)
}
\description{
Vector with 100 examples of DOI from papers. Is used as an example to build
an abstracts class.
}
\examples{
data(example_doi)
}
\keyword{datasets}
