% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_plus.R
\docType{data}
\name{example_plus}
\alias{example_plus}
\title{Example of plus model fitted with abstracts.}
\format{
An object of class \code{"plus"};
}
\usage{
data(example_plus)
}
\description{
Example of plus model fitted with abstracts.
}
\examples{
data(example_plus)
}
\keyword{datasets}
