% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_test.R
\docType{data}
\name{example_test}
\alias{example_test}
\title{Example of test data. Is a sample of the lacs dataset.}
\format{
An object of class \code{"abstacts"};
}
\usage{
data(example_test)
}
\description{
Example of test data. Is a sample of the lacs dataset.
}
\examples{
data(example_test)
}
\keyword{datasets}
