% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_title.R
\docType{data}
\name{example_title}
\alias{example_title}
\title{Vector with 100 examples of titles from papers. Is used as an example to build
an abstracts class.}
\format{
An object of class \code{"character"} with 100 titles
}
\usage{
data(example_title)
}
\description{
Vector with 100 examples of titles from papers. Is used as an example to build
an abstracts class.
}
\examples{
data(example_title)
}
\keyword{datasets}
