% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_train.R
\docType{data}
\name{example_train}
\alias{example_train}
\title{Example of train data. Is a sample of the lacs dataset.}
\format{
An object of class \code{"abstacts"};
}
\usage{
data(example_train)
}
\description{
Example of train data. Is a sample of the lacs dataset.
}
\examples{
data(example_train)
}
\keyword{datasets}
