% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_vocabulary.R
\docType{data}
\name{example_vocabulary}
\alias{example_vocabulary}
\title{Example of vocabulary. Is a vocabulary object used to test
the new_plus function.}
\format{
An object of class \code{"text2vec_vocabulary"};
}
\usage{
data(example_vocabulary)
}
\description{
Example of vocabulary. Is a vocabulary object used to test
the new_plus function.
}
\examples{
data(example_vocabulary)
}
\keyword{datasets}
