% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dtm.R
\name{get_dtm}
\alias{get_dtm}
\title{Get a document term matrix from abstracts}
\usage{
get_dtm(abstracts, vocabulary, tf_idf = TRUE)
}
\arguments{
\item{abstracts}{Could be an object of class abstracts (created via \code{get_abstracts} function) or
a \code{character}vector with abstracts}

\item{vocabulary}{A vocabulary created with the \code{get_vocabulary} function, from
an \code{abstracts} object}

\item{tf_idf}{logical. If it is true returns dtm weighted with tf idf. Otherwise return
binary dtm matrix}
}
\value{
Returns a document term matrix. Each row is a document, each term is a matrix from
vocabulary
}
\description{
Get a document term matrix from abstracts
}
\examples{
abstracts <- get_abstracts(lacsSample)
vocabulary <- get_vocabulary(abstracts, term_count_min = 10)
get_dtm(abstracts, vocabulary, tf_idf = FALSE)
}
