% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vocabulary.R
\name{get_vocabulary}
\alias{get_vocabulary}
\title{Create vocabulary from abstracts}
\usage{
get_vocabulary(abstracts, term_count = 2)
}
\arguments{
\item{abstracts}{An abstracts object build with get_abstracts function}

\item{term_count}{The minimum number of counts for an specific term in vocabulary}
}
\value{
A data.table with the vocabulary pruned by the term_count
}
\description{
Create vocabulary from abstracts
}
\examples{
abstracts <- get_abstracts(lacsSample)
v <- get_vocabulary(abstracts)
}
