% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rodents.R
\docType{data}
\name{na_rodents}
\alias{na_rodents}
\title{Data frame with 3628 abstracts with host-parasite information of
rodents of North America.
The data was manually curated by Dr. Beckett Sterner (ASU) and his team.
Is primarily used to test a lacs model.}
\format{
An object of class \code{"data.frame"};
}
\usage{
data(na_rodents)
}
\description{
Data frame with 3628 abstracts with host-parasite information of
rodents of North America.
The data was manually curated by Dr. Beckett Sterner (ASU) and his team.
Is primarily used to test a lacs model.
}
\examples{
data(na_rodents)
}
\keyword{datasets}
