% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_abstracts.R
\name{new_abstracts}
\alias{new_abstracts}
\title{Constructor function for the class abstracts
A constructor for the abstracts class}
\usage{
new_abstracts(
  doi = character(),
  title = character(),
  abstract = character(),
  class = character()
)
}
\arguments{
\item{doi}{a vector with DOI for each paper}

\item{title}{A vector of titles for each paper}

\item{abstract}{A vector of abstracts for each paper}

\item{class}{A vector with the class for each paper}
}
\value{
an object of class abstracts. Natively is a data.frame
}
\description{
Constructor function for the class abstracts
A constructor for the abstracts class
}
\examples{
new_abstracts(
doi = example_doi,
title = example_title,
abstract = example_abstract,
class = example_class
)
}
