% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_lacs.R
\name{new_lacs}
\alias{new_lacs}
\title{Constructor function for the class lacs.
A constructor for the lacs class}
\usage{
new_lacs(
  plus = plus::new_plus(),
  vocabulary = data.table::data.table(),
  train = data.frame(),
  call = character()
)
}
\arguments{
\item{plus}{An object of the class plus. Is the fit of the model}

\item{vocabulary}{An object of the class \code{text2vec_vocabulary}.}

\item{train}{An abstracts sample dataset used to train a plus model.}

\item{call}{The call of the function}
}
\value{
an object of class lacs.
}
\description{
Constructor function for the class lacs.
A constructor for the lacs class
}
\examples{
new_lacs(
plus = example_plus,
vocabulary = example_vocabulary,
train = example_train,
call = character()
)
}
