% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lacs.R
\name{predict}
\alias{predict}
\alias{predict.lacs}
\title{Predicted classification with a lacs model}
\usage{
predict(object, ...)

\method{predict}{lacs}(object, ..., abstracts, type = "class", plus = TRUE)
}
\arguments{
\item{object}{An object of lacs class to predict the text classification}

\item{\dots}{additional arguments}

\item{abstracts}{An object of class abstracts. Is a \code{data.frame} with
abstracts}

\item{type}{The type of the predicted response given a lacs model. Could be
the class (Binary output) or the probability given the plus model inside
the lacs object}

\item{plus}{If \code{TRUE} use the cutoff threshold to predict the class from
the plus object. If \code{FALSE} use the \code{cv.glmnet} object without
cutoff threshold.}
}
\value{
A vector of response. If the type is set as class returns a logical
  vector. If the type is response return a vector of probabilities.
}
\description{
Predicted classification with a lacs model
}
\examples{
model <- lacs(lacsSample[1:200, ])
lacsSample <- get_abstracts(lacsSample)
predict(model, abstracts = lacsSample )
predict(model, abstracts = lacsSample,  plus = FALSE )
predict(model, abstracts = lacsSample, type = "response")
}
