% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.lacs.R
\name{predictions}
\alias{predictions}
\alias{predictions.lacs}
\title{Title}
\usage{
predictions(object, ...)

\method{predictions}{lacs}(object, ..., abstracts, plus = TRUE)
}
\arguments{
\item{object}{A lacs object}

\item{\dots}{additional arguments}

\item{abstracts}{An object of class abstracts. Is a \code{data.frame} with
abstracts}

\item{plus}{Logical. If it is true use the cutoff threshold from the plus model.
This tends to reduce the false negatives and increase the recall. Othewise
gets the predicted value from glmnet default predicted response.}
}
\value{
A \code{data.frame} containing four columns. The truth independent
value (observed y), the probability to belong in Class1, the probability to
belong Class2 and the predicted class either the plus (use cutoff threshold)
or glmnet model.
}
\description{
Title
}
\examples{
model <- lacs(lacsSample[1:200, ])
lacsSample <- get_abstracts(lacsSample)
predictions(model, abstracts = lacsSample)
}
