% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.abstracts.R
\name{print.abstracts}
\alias{print.abstracts}
\title{Print an abstracts object}
\usage{
\method{print}{abstracts}(x, ...)
}
\arguments{
\item{x}{Object of the abstracts type to print.}

\item{...}{Passed on to [tbl_format_setup()].#'}
}
\value{
An object with a `print()` method that will print the input
  similarly to a tibble.
}
\description{
Print an abstracts object
}
\examples{
print(get_abstracts(lacsSample))
}
\keyword{internal}
