% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_plus.R
\name{fit_plus}
\alias{fit_plus}
\title{Fit a plus model to an abstracts object}
\usage{
fit_plus(
  abstracts,
  vocabulary = NULL,
  tf_idf = TRUE,
  alpha = 1,
  sample_use_time = 30,
  learning_rate = 1,
  qq = 0.1
)
}
\arguments{
\item{abstracts}{An abstracts object.}

\item{vocabulary}{A \code{ext2vec_vocabulary} object from a specific corpus
to create a document -  term matrix.}

\item{tf_idf}{Logical. Default is true, use a tf_idf in the document - term matrix S}

\item{alpha}{Plus parameter. The elastic net mixing parameter, with \eqn{0\le\alpha\le 1}}

\item{sample_use_time}{Plus parameter. Number of use time.}

\item{learning_rate}{Plus parameter. Learning rate. Default is 1.}

\item{qq}{Plus parameter. quantile threshold}
}
\value{
Return a plus object.(Positive and unlabeled Learning from Unbalanced
cases and Sparse structures, PLUS).
This is a cross-validated glmnet logistics regression optimized for positive and
unlabeled learning. This classification is applied to a dtm matrix created
with the vocabulary and the abstracts. The independent variables are the positive
and unknown classes of the provided \code{abstract} object.
}
\description{
Fit a plus model to an abstracts object
}
\examples{
abstracts <- get_abstracts(lacsSample)
v <- get_vocabulary(abstracts, term_count = 2)
s <- sample(seq(nrow(abstracts)), 65, replace = FALSE)
train <- abstracts[s, ]
model <- fit_plus(abstracts = train, vocabulary = v)
model
}
