% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abstracts.R
\name{get_abstracts}
\alias{get_abstracts}
\title{abstracts model helper function}
\usage{
get_abstracts(x = NULL)
}
\arguments{
\item{x}{A data frame with abstracts. Should contain doi, title, abstracts and a class column.
The class should be possitive or unknown.}
}
\value{
and object of the class abstracts
}
\description{
abstracts model helper function
}
\examples{
abstracts <- get_abstracts(lacsSample)
}
