% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vocabulary.R
\name{get_vocabulary}
\alias{get_vocabulary}
\title{Create vocabulary from abstracts}
\usage{
get_vocabulary(
  abstracts,
  term_count_min = 2,
  doc_proportion_min = 0,
  doc_proportion_max = 1
)
}
\arguments{
\item{abstracts}{An abstracts object build with get_abstracts function}

\item{term_count_min}{The minimum number of counts for an specific term in vocabulary}

\item{doc_proportion_min}{The minimum proportion of documents which should contain term}

\item{doc_proportion_max}{The maximum proportion of documents which should contain term.}
}
\value{
A data.table with the vocabulary pruned by the term_count
}
\description{
Create vocabulary from abstracts
}
\examples{
abstracts <- get_abstracts(lacsSample)
v <- get_vocabulary(abstracts, term_count_min = 10)
}
