% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_y.R
\name{get_y}
\alias{get_y}
\title{Get the independent variable for an abstracts object}
\usage{
get_y(abstracts)
}
\arguments{
\item{abstracts}{An object of class abstracts is a data frame
with a class column. Shoiuld have possitive and unknown classes}
}
\value{
A vector of binary output
}
\description{
Get the independent variable for an abstracts object
}
\examples{
abstracts <- get_abstracts(lacsSample)
get_y(abstracts)
}
