% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacs.R
\name{lacs}
\alias{lacs}
\title{Create an object of class lacs}
\usage{
lacs(
  abstracts,
  vocabulary = NULL,
  term_count_min = 2,
  doc_proportion_min = 0,
  doc_proportion_max = 1,
  alpha = 1,
  sample_use_time = 30,
  learning_rate = 1,
  qq = 0.1
)
}
\arguments{
\item{abstracts}{Object of \code{abstracts} class. A dataset of abstracts}

\item{vocabulary}{A vocabulary created with the \code{get_vocabulary} function, from
an \code{abstracts} object}

\item{term_count_min}{Vocabulary parameter. The minimum number of counts for an specific term in vocabulary}

\item{doc_proportion_min}{Vocabulary parameter. The minimum proportion of documents which should contain term}

\item{doc_proportion_max}{The Vocabulary parameter. The maximum proportion of documents which should contain term.}

\item{alpha}{Plus parameter. The elastic net mixing parameter, with \eqn{0\le\alpha\le 1}}

\item{sample_use_time}{Plus parameter. Number of use time.}

\item{learning_rate}{Plus parameter. Learning rate. Default is 1.}

\item{qq}{Plus parameter. quantile threshold}
}
\value{
An object of class lacs
}
\description{
Create an object of class lacs
}
\examples{
lacs(lacsSample, tail(example_vocabulary, 500))
}
