% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_abstracts.R
\name{validate_abstracts}
\alias{validate_abstracts}
\title{Validate a data frame with abstracts to construct an object of class abstracts via new_abstracts function}
\usage{
validate_abstracts(abstracts_df)
}
\arguments{
\item{abstracts_df}{A data frame with abstracts.
Should contain doi, title, abstracts and class columns in order to be valid.}
}
\value{
A validated \code{data.frame} with abstracts to construct an abstracts object
}
\description{
Validate a data frame with abstracts to construct an object of class abstracts via new_abstracts function
}
\examples{
validate_abstracts(lacsSample)
}
